+++ CHANGELOG.rst
@@ -25,6 +25,7 @@ Fixed
 ^^^^^
 - Missing a factor 0.5 in :func:`~tenpy.linalg.random_matrix.GUE`.
 - Allow :class:`~tenpy.networks.terms.TermList` to have terms with multiple operators acting on the same site.
+- Allow MPS indices outside unit cell in :meth:`~tenpy.models.lattice.Lattice.mps2lat_idx` and :meth:`~tenpy.models.lattice.Lattice.lat2mps_idx`.
 
 
 [0.4.0] - 2019-04-28